Fix: 101-print_number.c

+++ 0x04-more_functions_nested_loops/101-print_number.c
@@ -6,7 +6,7 @@
  */
 void print_number(int n)
 {
-	long long m = n;
+	long int m = n;
 
 	if (m < 0)
 	{
